added user and admin restrictions

+++ base/views.py
@@ -1,10 +1,13 @@
 from django.shortcuts import render,redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room,Topic
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
 from django.db.models import Q
 from .forms import RoomForm
+
 '''
 rooms=[
     {'id':1,'name':'Lets learn python'},
@@ -14,17 +17,31 @@ rooms=[
 '''
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')
         try:
-            user=User.objects.get(usename=username)
+            user=User.objects.get(username=username)
         except:
             messages.error(request,'User does not exist')
+        user=authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or password does not exist')
 
 
     context={}
     return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q') != None else ''
@@ -44,6 +61,7 @@ def room(requset,pk):
     context={'room': room}
     return render(requset,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -54,21 +72,28 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
-def updateRoom(request,pk):
-    room=Room.objects.get(id=pk)
-    form=RoomForm(instance=room)
+@login_required(login_url='login')
+def updateRoom(request, pk):
+    room = Room.objects.get(id=pk)
+    form = RoomForm(instance=room)
 
-    if request.method=='POST':
-        form=RoomForm(request.POST,instance=room)
+    # Check if the Room model has a 'user' attribute
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!!')
+
+    if request.method == 'POST':
+        form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
             return redirect('home')
-    
-    context={'form':form}
-    return render(request,'base/room_form.html',context)
+
+    context = {'form': form}
+    return render(request, 'base/room_form.html', context)
         
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!!')
     if request.method=='POST':
         room.delete()
         return redirect('home')
